Refactor filter state and book fetching to Sidebar component
