+++ src/pages/books.js
@@ -1,5 +1,3 @@
-import { supabase } from "@/utils/supabase";
-import { useEffect } from "react";
 import { useState } from "react";
 import Book from "@/components/Book";
 import Sidebar from "@/components/Sidebar";
@@ -7,37 +5,9 @@ import Sidebar from "@/components/Sidebar";
 export default function Books() {
     const [books, setBooks] = useState([]);
     const [browsing, setBrowsing] = useState(false);
-    const [filters, setFilters] = useState([true, false]);
-
-    useEffect(() => {
-        fetchBooks(filters);
-    }, [filters]);
-
-    let fetchBooks = async (filters) => {
-        let data;
-        if (filters[0] && !filters[1]) {
-            console.log("TEST");
-            data = await supabase
-                .from("book_data2")
-                .select()
-                .eq("Exclusive Shelf", "read");
-        } else if (filters[0] && filters[1]) {
-            data = await supabase.from("book_data2").select();
-        } else if (!filters[0] && filters[1]) {
-            data = await supabase
-                .from("book_data2")
-                .select()
-                .eq("Exclusive Shelf", "to-read");
-        }
-        setBooks(data?.data);
-    };
 
     let updateBooks = (book_list) => {
         setBooks(book_list);
-    };
-
-    let updateFilters = (filters) => {
-        setFilters(filters);
     };
 
     let handleSearchClick = () => {
@@ -63,10 +33,8 @@ export default function Books() {
             <div className="flex">
                 <Sidebar
                     updateBooks={updateBooks}
-                    updateFilters={updateFilters}
                     browsing={browsing}
                     books={books}
-                    filters={filters}
                 />
                 <div
                     className={`${
